+++ src/main/java/org/prgrms/kdt/io/ConsoleIO.java
@@ -1,17 +1,23 @@
 package org.prgrms.kdt.io;
 
+import org.prgrms.kdt.voucher.Voucher;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class ConsoleIO {
-    private final Input input;
-    private final Output output;
-
-    public static final String COMMAND_LIST_PROMPT = "=== Voucher Program ===" + System.lineSeparator() +
+    public static final String COMMAND_LIST_PROMPT = System.lineSeparator() +
+            "=== Voucher Program ===" + System.lineSeparator() +
             "Type exit to exit the program." + System.lineSeparator() +
             "Type create to create a new voucher." + System.lineSeparator() +
             "Type list to list all vouchers" + System.lineSeparator();
-    public static final String COMMAND_ERROR_PROMPT = "지원하지 않는 명령어입니다." + System.lineSeparator();
+    public static final String VOUCHER_TYPE_PROMPT = System.lineSeparator() + "Enter voucher type. (fixed or percent)";
+    public static final String VOUCHER_AMOUNT_PROMPT = "Enter a value. (fixed - 0~ | percent - 0~100)";
+
+    private final Input input;
+    private final Output output;
 
     public ConsoleIO(Input input, Output output) {
         this.input = input;
@@ -20,23 +26,31 @@ public class ConsoleIO {
 
     public String getCommand() {
         output.printText(COMMAND_LIST_PROMPT);
+
         return input.inputText();
     }
 
-    public void printCommandError() {
-        output.printText(COMMAND_ERROR_PROMPT);
-    }
+    public String getType() {
+        output.printText(VOUCHER_TYPE_PROMPT);
 
-    public String getInput() {
         return input.inputText();
-    }
-
-
-    public String getVoucherType() {
-        return null;
     }
 
     public String getVoucherAmount() {
-        return null;
+        output.printText(VOUCHER_AMOUNT_PROMPT);
+
+        return input.inputText();
+    }
+
+    public void printError(String message) {
+        output.printText(message);
+    }
+
+    public void printVouchers(List<Voucher> vouchers) {
+        output.printText(
+                vouchers.stream()
+                        .map(Voucher::toString)
+                        .collect(Collectors.joining("\n"))
+        );
     }
 }
