--- src/main/java/org/prgrms/kdt/voucher/VoucherInfo.java
@@ -1,4 +1,0 @@
-package org.prgrms.kdt.voucher;
-
-public class VoucherInfo {
-}
