+++ src/main/java/org/prgrms/kdt/voucher/VoucherMapper.java
@@ -5,7 +5,18 @@ import org.springframework.stereotype.Component;
 @Component
 public class VoucherMapper {
 
-    public Voucher mapFrom(VoucherInfo voucherInfo) {
-        return new Voucher();
+    public Voucher MetaDataToVoucher(VoucherMetaData voucherMetaData) {
+        return new Voucher(
+                toVoucherType(voucherMetaData.getType()),
+                toVoucherAmount(voucherMetaData.getAmount())
+        );
+    }
+
+    private VoucherAmount toVoucherAmount(String amount) {
+        return new VoucherAmount(amount);
+    }
+
+    private VoucherType toVoucherType(String type) {
+        return VoucherType.findVoucherType(type);
     }
 }
